Implement bell
Bell is now correctly pawn of Player not the other way around. Player
updates bell position if it has it equipped.
Still, the first bell is an atribute of PlayerObject from beginning,
which is less than ideal perhaps, making PlayerObject more bloated
than it needs to be but such is life.